requirements.txt
langchain
openai
faiss-cpu
python-dotenv
tiktoken
pysrt
fastapi